split dbgfuncCoprocessor
