+++ dbms/src/Server/MockComputeClient.h
@@ -11,6 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+#pragma once
 
 #include <Common/Exception.h>
 #include <Common/grpcpp.h>
