+++ dbms/src/TestUtils/MPPTaskTestUtils.cpp
@@ -11,8 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#include <Debug/dbgFuncCoprocessor.h>
-#include <Debug/dbgFuncCoprocessorUtils.h>
 #include <Server/MockComputeClient.h>
 #include <TestUtils/MPPTaskTestUtils.h>
