+++ dbms/src/TestUtils/mockExecutor.h
@@ -17,7 +17,7 @@
 #include <Core/ColumnsWithTypeAndName.h>
 #include <Debug/MockExecutor/AstToPB.h>
 #include <Debug/MockStorage.h>
-#include <Debug/dbgFuncCoprocessor.h>
+#include <Debug/dbgQueryCompiler.h>
 #include <Interpreters/Context.h>
 #include <Parsers/ASTFunction.h>
 #include <Storages/Transaction/Collator.h>
